Changed Environment.TickCount to DateTime.Ticks to avoid Win10 FastBoot issues

+++ DCS-SR-Client/Network/Models/RadioSendingState.cs
@@ -5,7 +5,7 @@ namespace Ciribob.DCS.SimpleRadio.Standalone.Client.Network
     public class RadioSendingState
     {
         [JsonIgnore]
-        public double LastSentAt { get; set; }
+        public long LastSentAt { get; set; }
 
         public bool IsSending { get; set; }
